modified files

+++ hashtags_for_locations/analysis_predict_hashtags_for_locations.py
@@ -378,7 +378,7 @@ class PredictHashtagsForLocationsPlots():
             _, num_of_items = zip(*ltuo_bucket_id_and_num_of_items)
             total_num_of_items = sum(num_of_items)+0.0
             for bucket_id, num_of_items in ltuo_bucket_id_and_num_of_items:
-                mf_bucket_id_to_num_of_items[bucket_id]+=int(100*(num_of_items/total_num_of_items))
+                mf_bucket_id_to_num_of_items[bucket_id]+=100*(num_of_items/total_num_of_items)
             print mf_bucket_id_to_num_of_items
             print sum(mf_bucket_id_to_num_of_items.values())
             exit()
